Convert to Bikeshed, step 1
Convert the body of the document, and make use of bikeshed's facilities
for simplified markup, cross references, auto links, ToC generation,
index generation…

This includes numerous markup fixups, both to take advantage of bikshed'
capacities, to improve consistency, and to make the whole thing easier
to work with.

The text of the process itself is unchanged.

Due the extensive source reformatting, a source-level diff is likely to
be quite unusable. However, using a tool like
https://services.w3.org/htmldiff can show that the text is changed and
unharmed.